+++ CHANGELOG.md
@@ -20,7 +20,7 @@
 - (Improvement) Remove PodSchedulingFailure condition instead of setting to false, restart pod if it could not be scheduled
 - (Feature) Add ArangoMember overrides
 - (Feature) ArangoMember Removal Priority
-- (Feature) Add --deployment.feature.init-containers-copy-limits (default enabled)
+- (Feature) Add --deployment.feature.init-containers-copy-resources (default enabled)
 
 ## [1.2.32](https://github.com/arangodb/kube-arangodb/tree/1.2.32) (2023-08-07)
 - (Feature) Backup lifetime - remove Backup once its lifetime has been reached